apps/api/pyproject.toml
[project]
name = "agent-core-api"
version = "0.1.0"
description = "Agent Core backend for MiMo-integrated agent workflows"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "redis>=5.2.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"